Refactoring vite.config.
