+++ faster-than-realtime/client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test-js",
+  "name": "faster-than-realtime",
   "private": true,
   "version": "0.0.0",
   "type": "module",
